requirements.txt
Django==2.2.1
django-crispy-forms==1.7.2
django-rest-framework==0.1.0
djangorestframework==3.9.3
python-dateutil==2.8.0
requests==2.9.1
